Separate tests to different suits.
