+++ HomeWork_16/src/test/java/HomeWorks/CommonClass.java
@@ -2,6 +2,8 @@ package HomeWorks;
 
 import HW_13_Java_8_Separate_Methods.MainRunner;
 import HW_13_Java_8_Separate_Methods.User;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 
@@ -9,6 +11,8 @@ import java.util.List;
 
 public class CommonClass {
     protected List<User> users;
+
+    protected static final Logger logger = LogManager.getLogger(PositiveTestRunner.class);
 
     @BeforeMethod
     public void FillInUsersList() {
